refactor ssd1681 to inherit from ssd168x base class
