+++ devices/ePaper/Drivers/Ssd168x/Ssd1681/Ssd1681.cs
@@ -16,7 +16,7 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
     /// <summary>
     /// A driver class for the SSD1681 display controller.
     /// </summary>
-    public sealed class Ssd1681 : IEPaperDisplay
+    public sealed class Ssd1681 : Ssd168x
     {
         /// <summary>
         /// The max supported clock frequency for the SSD1681 controller. 20MHz.
@@ -28,26 +28,7 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
         /// </summary>
         public const SpiMode SpiMode = System.Device.Spi.SpiMode.Mode0;
 
-        private const int PagesPerFrame = 5;
-        private const int FirstPageIndex = 0;
-
-        private readonly bool _shouldDispose;
-
-        private SpiDevice _spiDevice;
-        private GpioController _gpioController;
-        private GpioPin _resetPin;
-        private GpioPin _busyPin;
-        private GpioPin _dataCommandPin;
-
-        private int _currentFrameBufferPage;
-        private int _currentFrameBufferPageLowerBound;
-        private int _currentFrameBufferPageUpperBound;
-        private int _currentFrameBufferStartXPosition;
-        private int _currentFrameBufferStartYPosition;
-
         private FrameBuffer2BitPerPixel _frameBuffer2bpp;
-
-        private bool _disposed;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="Ssd1681"/> class.
@@ -61,6 +42,7 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
         /// <param name="gpioController">The <see cref="GpioController"/> to use when initializing the pins.</param>
         /// <param name="enableFramePaging">Page the frame buffer and all operations to use less memory.</param>
         /// <param name="shouldDispose"><see langword="true"/> to dispose of the <see cref="GpioController"/>.</param>
+        /// <exception cref="ArgumentNullException"><paramref name="spiDevice"/> is null.</exception>
         /// <exception cref="ArgumentOutOfRangeException">Display width and height can't be less than 0 or greater than 200.</exception>
         /// <remarks>
         /// For a 200x200 SSD1681 display, a full Frame requires about 5KB of RAM ((200 * 200) / 8). SSD1681 has 2 RAMs for B/W and Red pixel.
@@ -76,88 +58,32 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
             int height,
             GpioController gpioController = null,
             bool enableFramePaging = true,
-            bool shouldDispose = true)
+            bool shouldDispose = true) : base(spiDevice, resetPin, busyPin, dataCommandPin, width, height, gpioController, enableFramePaging, shouldDispose)
         {
-            if (width <= 0 || width > 200)
-            {
-                throw new ArgumentOutOfRangeException(nameof(width));
-            }
-
-            if (height <= 0 || height > 200)
-            {
-                throw new ArgumentOutOfRangeException(nameof(height));
-            }
-
-            _spiDevice = spiDevice ?? throw new ArgumentNullException(nameof(spiDevice));
-            _gpioController = gpioController ?? new GpioController();
-
-            // setup the gpio pins
-            _resetPin = resetPin >= 0 ? gpioController.OpenPin(resetPin, PinMode.Output) : null;
-            _dataCommandPin = gpioController.OpenPin(dataCommandPin, PinMode.Output);
-            _busyPin = gpioController.OpenPin(busyPin, PinMode.Input);
-
-            _shouldDispose = shouldDispose || gpioController == null;
-
-            Width = width;
-            Height = height;
-            PagedFrameDrawEnabled = enableFramePaging;
-            PowerState = PowerState.Unknown;
-
-            InitializeFrameBuffer(width, height, enableFramePaging);
-            CalculateFrameBufferPageBounds();
         }
 
         /// <inheritdoc/>
-        public int Width { get; }
-
-        /// <inheritdoc/>
-        public int Height { get; }
-
-        /// <inheritdoc/>
-        public IFrameBuffer FrameBuffer
+        public override IFrameBuffer FrameBuffer
         {
             get
             {
                 return _frameBuffer2bpp;
             }
-        }
 
-        /// <summary>
-        /// Gets the current power state of the display panel.
-        /// </summary>
-        public PowerState PowerState { get; private set; }
+            protected set
+            {
+                _frameBuffer2bpp = (FrameBuffer2BitPerPixel)value;
+            }
+        }
 
         /// <inheritdoc/>
-        public bool PagedFrameDrawEnabled { get; }
-
-        /// <summary>
-        /// Performs the required steps to "power on" the display.
-        /// </summary>
-        public void PowerOn()
-        {
-            HardwareReset();
-            SoftwareReset();
-
-            PowerState = PowerState.PoweredOn;
-        }
-
-        /// <summary>
-        /// Puts the display to sleep using the specified <see cref="SleepMode"/>.
-        /// </summary>
-        /// <param name="sleepMode">The sleep mode to use when powering down the display.</param>
-        public void PowerDown(SleepMode sleepMode = SleepMode.Normal)
-        {
-            SendCommand((byte)Command.DeepSleepMode);
-            SendData((byte)sleepMode);
-
-            PowerState = PowerState.PoweredOff;
-        }
+        protected override int PagesPerFrame { get; } = 4;
 
         /// <summary>
         /// Perform the required initialization steps to set up the display.
         /// </summary>
         /// <exception cref="InvalidOperationException">Unable to initialize the display until it has been powered on. Call PowerOn() first.</exception>
-        public void Initialize()
+        public override void Initialize()
         {
             if (PowerState != PowerState.PoweredOn)
             {
@@ -201,91 +127,18 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
         }
 
         /// <inheritdoc/>
-        public void PerformFullRefresh()
-        {
-            SendCommand((byte)Command.BoosterSoftStartControl);
-            SendData(0x8b, 0x9c, 0x96, 0x0f);
-
-            SendCommand((byte)Command.DisplayUpdateControl2);
-            SendData((byte)RefreshMode.FullRefresh); // Display Mode 1 (Full Refresh)
-
-            SendCommand((byte)Command.MasterActivation);
-            WaitReady();
-        }
-
-        /// <inheritdoc/>
-        public void PerformPartialRefresh()
-        {
-            SendCommand((byte)Command.BoosterSoftStartControl);
-            SendData(0x8b, 0x9c, 0x96, 0x0f);
-
-            SendCommand((byte)Command.DisplayUpdateControl2);
-            SendData((byte)RefreshMode.PartialRefresh); // Display Mode 2 (Partial Refresh)
-
-            SendCommand((byte)Command.MasterActivation);
-            WaitReady();
-        }
-
-        /// <inheritdoc/>
-        public void Clear(bool triggerPageRefresh = false)
-        {
-            SetFrameBufferPage(FirstPageIndex);
-
-            // paging is enabled, flush as per number of pages to ensure all display RAM is cleared
-            if (PagedFrameDrawEnabled)
-            {
-                do
-                {
-                    Flush();
-
-                    // sleep for 20ms to allow other threads to have their chance to execute
-                    Thread.Sleep(20);
-
-                    _currentFrameBufferPage++;
-                    CalculateFrameBufferPageBounds();
-                }
-                while (_currentFrameBufferPage < PagesPerFrame);
-
-                _currentFrameBufferPage = FirstPageIndex;
-                CalculateFrameBufferPageBounds();
-            }
-            else
-            {
-                // paging is disabled, so the internal frame covers the entire display frame. we only need to flush once.
-                Flush();
-            }
-
-            if (triggerPageRefresh)
-            {
-                PerformFullRefresh();
-            }
-        }
-
-        /// <inheritdoc/>
-        public void Flush()
+        public override void Flush()
         {
             // write B/W and Color (RED) frame to the display's RAM.
             DirectDrawBuffer(
-                _currentFrameBufferStartXPosition,
-                _currentFrameBufferStartYPosition,
+                CurrentFrameBufferStartXPosition,
+                CurrentFrameBufferStartYPosition,
                 _frameBuffer2bpp.BlackBuffer.Buffer);
 
             DirectDrawColorBuffer(
-                _currentFrameBufferStartXPosition,
-                _currentFrameBufferStartYPosition,
+                CurrentFrameBufferStartXPosition,
+                CurrentFrameBufferStartYPosition,
                 _frameBuffer2bpp.ColorBuffer.Buffer);
-        }
-
-        /// <inheritdoc/>
-        public void SetPosition(int x, int y)
-        {
-            EnforceBounds(ref x, ref y);
-
-            SendCommand((byte)Command.SetRAMAddressCounterX);
-            SendData((byte)(x / 8)); // each row can have up to 200 points each represented by a single bit.
-
-            SendCommand((byte)Command.SetRAMAddressCounterY);
-            SendData((byte)y);
         }
 
         /// <summary>
@@ -318,7 +171,7 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
         /// if R == 0 and G == 0 and B == 0 then write a black pixel to B/W Buffer/RAM
         /// else, assume white pixel and write to B/W Buffer/RAM.
         /// </remarks>
-        public void DrawPixel(int x, int y, Color color)
+        public override void DrawPixel(int x, int y, Color color)
         {
             // ignore out of bounds draw attempts
             if (x < 0 || x >= Width || y < 0 || y >= Height)
@@ -327,11 +180,11 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
             }
 
             var frameByteIndex = GetFrameBufferIndex(x, y);
-            var pageByteIndex = frameByteIndex - _currentFrameBufferPageLowerBound;
+            var pageByteIndex = frameByteIndex - CurrentFrameBufferPageLowerBound;
 
             // if the specified point falls in the current page, update the buffer
-            if (_currentFrameBufferPageLowerBound <= frameByteIndex
-                && frameByteIndex < _currentFrameBufferPageUpperBound)
+            if (CurrentFrameBufferPageLowerBound <= frameByteIndex
+                && frameByteIndex < CurrentFrameBufferPageUpperBound)
             {
                 /*
                  * Lookup Table for colors on SSD1681
@@ -379,7 +232,7 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
 
         /// <summary>
         /// Draws the specified buffer directly to the Black/White RAM on the display.
-        /// Call <see cref="PerformFullRefresh"/> after to update the display.
+        /// Call <see cref="Ssd168x.PerformFullRefresh"/> after to update the display.
         /// </summary>
         /// <param name="startXPos">X start position.</param>
         /// <param name="startYPos">Y start position.</param>
@@ -394,7 +247,7 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
 
         /// <summary>
         /// Draws the specified buffer directly to the Red RAM on the display.
-        /// Call <see cref="PerformFullRefresh"/> after to update the display.
+        /// Call <see cref="Ssd168x.PerformFullRefresh"/> after to update the display.
         /// </summary>
         /// <param name="startXPos">X start position.</param>
         /// <param name="startYPos">Y start position.</param>
@@ -407,174 +260,13 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
             SendData(buffer);
         }
 
-        /// <inheritdoc/>
-        public void BeginFrameDraw()
-        {
-            // make sure we start from the first page with clear buffers
-            SetFrameBufferPage(FirstPageIndex);
-        }
-
-        /// <inheritdoc/>
-        public bool NextFramePage()
-        {
-            if (PagedFrameDrawEnabled && _currentFrameBufferPage < PagesPerFrame - 1)
-            {
-                Flush();
-
-                _currentFrameBufferPage++;
-                SetFrameBufferPage(_currentFrameBufferPage);
-
-                return true;
-            }
-
-            return false;
-        }
-
-        /// <inheritdoc/>
-        public void EndFrameDraw()
-        {
-            Flush();
-        }
-
-        /// <inheritdoc/>
-        public void SendCommand(params byte[] command)
-        {
-            // make sure we are setting data/command pin to low (command mode)
-            _dataCommandPin.Write(PinValue.Low);
-
-            foreach (var b in command)
-            {
-                // write the command byte to the display controller
-                _spiDevice.WriteByte(b);
-            }
-        }
-
-        /// <inheritdoc/>
-        public void SendData(params byte[] data)
-        {
-            // set the data/command pin to high to indicate to the display we will be sending data
-            _dataCommandPin.Write(PinValue.High);
-
-            foreach (var @byte in data)
-            {
-                _spiDevice.WriteByte(@byte);
-            }
-
-            // go back to low (command mode)
-            _dataCommandPin.Write(PinValue.Low);
-        }
-
-        /// <inheritdoc/>
-        public void WaitReady()
-        {
-            while (_busyPin.Read() == PinValue.High)
-            {
-                WaitMs(5);
-            }
-        }
-
-        /// <summary>
-        /// Performs the hardware reset commands sequence on the display.
-        /// </summary>
-        private void HardwareReset()
-        {
-            if (_resetPin == null)
-            {
-                // caller opted to reset outside of the driver by not passing a valid reset pin number.
-                // do nothing.
-                return;
-            }
-
-            // specs say to wait 10ms after supplying voltage to the display
-            WaitMs(10);
-
-            _resetPin.Write(PinValue.Low);
-            WaitMs(200);
-
-            _resetPin.Write(PinValue.High);
-            WaitMs(200);
-        }
-
-        /// <summary>
-        /// Performs the software reset commands sequence on the display.
-        /// </summary>
-        private void SoftwareReset()
-        {
-            SendCommand((byte)Command.SoftwareReset);
-
-            WaitReady();
-            WaitMs(10);
-        }
-
-        /// <summary>
-        /// A simple wait method that wraps <see cref="Thread.Sleep(int)"/>.
-        /// </summary>
-        /// <param name="milliseconds">Number of milliseconds to sleep.</param>
-        private void WaitMs(int milliseconds)
-        {
-            Thread.Sleep(milliseconds);
-        }
-
-        /// <summary>
-        /// Snaps the provided coordinates to the lower bounds of the display if out of allowed range.
-        /// </summary>
-        /// <param name="x">The X position.</param>
-        /// <param name="y">The Y position.</param>
-        private void EnforceBounds(ref int x, ref int y)
-        {
-            x = x < 0 || x > Width - 1 ? 0 : x;
-            y = y < 0 || y > Height - 1 ? 0 : y;
-        }
-
-        /// <summary>
-        /// Gets the index of the byte containing the pixel specified by the <paramref name="x"/> and <paramref name="y"/> parameters.
-        /// </summary>
-        /// <param name="x">The X position of the pixel.</param>
-        /// <param name="y">The Y position of the pixel.</param>
-        /// <returns>The index of the byte in the frame buffer which contains the specified pixel.</returns>
-        private int GetFrameBufferIndex(int x, int y)
-        {
-            // x specifies the column
-            return (x + (y * Width)) / 8;
-        }
-
-        /// <summary>
-        /// Gets the X position from a buffer index.
-        /// </summary>
-        /// <param name="index">The buffer index.</param>
-        /// <returns>The X position of a pixel.</returns>
-        private int GetXPositionFromFrameBufferIndex(int index)
-        {
-            if (index <= 0)
-            {
-                return 0;
-            }
-
-            return (index * 8) % Width;
-        }
-
-        /// <summary>
-        /// Gets the Y position from a buffer index.
-        /// </summary>
-        /// <param name="index">The buffer index.</param>
-        /// <returns>The Y position of a pixel.</returns>
-        private int GetYPositionFromFrameBufferIndex(int index)
-        {
-            if (index <= 0)
-            {
-                return 0;
-            }
-
-            return index * 8 / Height;
-        }
-
         /// <summary>
         /// Initializes a new instance of the intenral frame buffer. Supports paging.
         /// </summary>
         /// <param name="width">The width of the frame buffer in pixels.</param>
         /// <param name="height">The height of the frame buffer in pixels.</param>
         /// <param name="enableFramePaging">If <see langword="true"/>, enables paging the frame.</param>
-        private void InitializeFrameBuffer(int width, int height, bool enableFramePaging)
+        protected override void InitializeFrameBuffer(int width, int height, bool enableFramePaging)
         {
             var frameBufferHeight = enableFramePaging ? height / PagesPerFrame : height;
             _frameBuffer2bpp = new FrameBuffer2BitPerPixel(frameBufferHeight, width);
@@ -583,11 +275,11 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
         /// <summary>
         /// Sets the current active frame buffer page to the specified page index.
         /// Existing frame buffer is reused by clearing it first and page bounds are recalculated.
-        /// Make sure to call <see cref="PerformFullRefresh"/> or <see cref="PerformPartialRefresh"/>
+        /// Make sure to call <see cref="Ssd168x.PerformFullRefresh"/> or <see cref="Ssd168x.PerformPartialRefresh"/>
         /// to persist the frame to the display's RAM before calling this method.
         /// </summary>
         /// <param name="page">The frame buffer page index to move to.</param>
-        private void SetFrameBufferPage(int page)
+        protected override void SetFrameBufferPage(int page)
         {
             if (page < 0 || page >= PagesPerFrame)
             {
@@ -595,67 +287,9 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
             }
 
             _frameBuffer2bpp.Clear();
-            _currentFrameBufferPage = page;
+            CurrentFrameBufferPage = page;
 
             CalculateFrameBufferPageBounds();
         }
-
-        /// <summary>
-        /// Calculates the upper and lower bounds of the current frame buffer page.
-        /// </summary>
-        private void CalculateFrameBufferPageBounds()
-        {
-            // black and color buffers have the same size, so we will work with only the black buffer
-            // in these calculations.
-            _currentFrameBufferPageLowerBound = _currentFrameBufferPage * _frameBuffer2bpp.BlackBuffer.BufferByteCount;
-            _currentFrameBufferPageUpperBound = (_currentFrameBufferPage + 1) * _frameBuffer2bpp.BlackBuffer.BufferByteCount;
-
-            _currentFrameBufferStartXPosition = GetXPositionFromFrameBufferIndex(_currentFrameBufferPageLowerBound);
-            _currentFrameBufferStartYPosition = GetYPositionFromFrameBufferIndex(_currentFrameBufferPageLowerBound);
-
-            FrameBuffer.StartPoint = new Point(_currentFrameBufferStartXPosition, _currentFrameBufferStartYPosition);
-            FrameBuffer.CurrentFramePage = _currentFrameBufferPage;
-        }
-
-        #region IDisposable
-
-        private void Dispose(bool disposing)
-        {
-            if (!_disposed)
-            {
-                if (disposing)
-                {
-                    _resetPin?.Dispose();
-                    _resetPin = null;
-
-                    _busyPin?.Dispose();
-                    _busyPin = null;
-
-                    _dataCommandPin?.Dispose();
-                    _dataCommandPin = null;
-
-                    if (_shouldDispose)
-                    {
-                        _gpioController?.Dispose();
-                    }
-
-                    _spiDevice = null;
-                    _gpioController = null;
-                    _frameBuffer2bpp = null;
-                }
-
-                _disposed = true;
-            }
-        }
-
-        /// <inheritdoc />
-        public void Dispose()
-        {
-            // Do not change this code. Put cleanup code in 'Dispose(bool disposing)' method
-            Dispose(disposing: true);
-            GC.SuppressFinalize(this);
-        }
-
-        #endregion
     }
 }
