+++ devices/ePaper/Samples/SSD1681Sample/Program.cs
@@ -7,6 +7,7 @@ using System.Drawing;
 
 using Iot.Device.EPaper;
 using Iot.Device.EPaper.Buffers;
+using Iot.Device.EPaper.Drivers.Ssd168x;
 using Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681;
 using Iot.Device.EPaper.Enums;
 using Iot.Device.EPaper.Fonts;
